Basic GUI
Checks how well the current solution works with a GUI, very very basic(
and manual)

+++ Pacman/Bot.cpp
@@ -56,10 +56,10 @@ int Bot::getID()
 //
 //
 
-sf::Vector2i Bot::move()
+sf::Vector2f Bot::move()
 {
-	if (getMovementProgress() == 0 && // the bot hasn't started moving yet( prevents cycling the same knot)
-		map->getLogical(getY(), getX()) == gv::knotSquare) // square the bot is on is a 'knot'
+	if (getMovementProgress() == 0 && // if the bot hasn't started moving yet( prevents cycling the same knot)
+		map->getLogical(getY(), getX()) == gv::knotSquare) // if the square the bot is on is a 'knot'
 	{
 		pickRandomDirection();
 	}
